refactor: 홈화면 탭, 헤더 컴포넌트 분리
